fraud_service/requirements.txt
fastapi==0.112.0
uvicorn==0.30.3
confluent-kafka==2.5.0
cassandra-driver==3.29.2
pydantic==2.8.2
pydantic-settings==2.11.0
python-dotenv==1.0.1
requests==2.32.3
pyjwt==2.9.0
